% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_functions.R
\name{activities.list}
\alias{activities.list}
\title{Returns a list of activities visible to the current logged in user. Visible activities are determined by the visiblity settings of the object that was acted on, e.g. Drive files a user can see. An activity is a record of past events. Multiple events may be merged if they are similar. A request is scoped to activities from a given Google service using the source parameter.}
\usage{
activities.list(drive.ancestorId = NULL, drive.fileId = NULL,
  groupingStrategy = NULL, pageSize = NULL, pageToken = NULL,
  source = NULL, userId = NULL)
}
\arguments{
\item{drive.ancestorId}{Identifies the Drive folder containing the items for which to return activities}

\item{drive.fileId}{Identifies the Drive item to return activities for}

\item{groupingStrategy}{Indicates the strategy to use when grouping singleEvents items in the associated combinedEvent object}

\item{pageSize}{The maximum number of events to return on a page}

\item{pageToken}{A token to retrieve a specific page of results}

\item{source}{The Google service from which to return activities}

\item{userId}{Indicates the user to return activity for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/activity
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/activity, https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.metadata.readonly, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/activity/}{Google Documentation}
}

