% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{Activity}
\alias{Activity}
\title{Google Apps Activity API Objects 
Provides a historical view of activity.}
\usage{
Activity(combinedEvent = NULL, singleEvents = NULL)
}
\arguments{
\item{combinedEvent}{The fields common to all of the singleEvents that make up the Activity}

\item{singleEvents}{A list of all the Events that make up the Activity}
}
\value{
Activity object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:35:50
filename: /Users/mark/dev/R/autoGoogleAPI/googleappsactivityv1.auto/R/appsactivity_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Activity Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Activity resource is a combined view of multiple events. An activity has a list of individual events and a combined view of the common fields among all events.
}
\seealso{
Other res: \code{\link{Event}},
  \code{\link{ListActivitiesResponse}}, \code{\link{Move}},
  \code{\link{Parent}}, \code{\link{PermissionChange}},
  \code{\link{Permission}}, \code{\link{Photo}},
  \code{\link{Rename}}, \code{\link{Target}},
  \code{\link{User}}
}

