% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_functions.R
\docType{package}
\name{appsactivity_googleAuthR}
\alias{appsactivity_googleAuthR}
\alias{appsactivity_googleAuthR-package}
\title{G Suite Activity API
Provides a historical view of activity.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:28:12
filename: /Users/mark/dev/R/autoGoogleAPI/googleappsactivityv1.auto/R/appsactivity_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/activity
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.readonly
}
}
