% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{Event}
\alias{Event}
\title{Event Object}
\usage{
Event(additionalEventTypes = NULL, eventTimeMillis = NULL,
  fromUserDeletion = NULL, move = NULL, permissionChanges = NULL,
  primaryEventType = NULL, rename = NULL, target = NULL, user = NULL)
}
\arguments{
\item{additionalEventTypes}{Additional event types}

\item{eventTimeMillis}{The time at which the event occurred formatted as Unix time in milliseconds}

\item{fromUserDeletion}{Whether this event is caused by a user being deleted}

\item{move}{Extra information for move type events, such as changes in an object's parents}

\item{permissionChanges}{Extra information for permissionChange type events, such as the user or group the new permission applies to}

\item{primaryEventType}{The main type of event that occurred}

\item{rename}{Extra information for rename type events, such as the old and new names}

\item{target}{Information specific to the Target object modified by the event}

\item{user}{Represents the user responsible for the event}
}
\value{
Event object
}
\description{
Event Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the changes associated with an action taken by a user.
}
