% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{ListActivitiesResponse}
\alias{ListActivitiesResponse}
\title{ListActivitiesResponse Object}
\usage{
ListActivitiesResponse(activities = NULL, nextPageToken = NULL)
}
\arguments{
\item{activities}{List of activities}

\item{nextPageToken}{Token for the next page of results}
}
\value{
ListActivitiesResponse object
}
\description{
ListActivitiesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response from the list request. Contains a list of activities and a token to retrieve the next page of results.
}
\seealso{
Other res: \code{\link{Activity}}, \code{\link{Event}},
  \code{\link{Move}}, \code{\link{Parent}},
  \code{\link{PermissionChange}}, \code{\link{Permission}},
  \code{\link{Photo}}, \code{\link{Rename}},
  \code{\link{Target}}, \code{\link{User}}
}

