% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{Move}
\alias{Move}
\title{Move Object}
\usage{
Move(addedParents = NULL, removedParents = NULL)
}
\arguments{
\item{addedParents}{The added parent(s)}

\item{removedParents}{The removed parent(s)}
}
\value{
Move object
}
\description{
Move Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about changes in an object's parents as a result of a move type event.
}
\seealso{
Other res: \code{\link{Activity}}, \code{\link{Event}},
  \code{\link{ListActivitiesResponse}},
  \code{\link{Parent}}, \code{\link{PermissionChange}},
  \code{\link{Permission}}, \code{\link{Photo}},
  \code{\link{Rename}}, \code{\link{Target}},
  \code{\link{User}}
}

