% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{Parent}
\alias{Parent}
\title{Parent Object}
\usage{
Parent(id = NULL, isRoot = NULL, title = NULL)
}
\arguments{
\item{id}{The parent's ID}

\item{isRoot}{Whether this is the root folder}

\item{title}{The parent's title}
}
\value{
Parent object
}
\description{
Parent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a parent object. For example, a folder in Drive is a parent for all files within it.
}
