% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{Permission}
\alias{Permission}
\title{Permission Object}
\usage{
Permission(name = NULL, permissionId = NULL, role = NULL, type = NULL,
  user = NULL, withLink = NULL)
}
\arguments{
\item{name}{The name of the user or group the permission applies to}

\item{permissionId}{The ID for this permission}

\item{role}{Indicates the Google Drive permissions role}

\item{type}{Indicates how widely permissions are granted}

\item{user}{The user's information if the type is USER}

\item{withLink}{Whether the permission requires a link to the file}
}
\value{
Permission object
}
\description{
Permission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about the permissions and type of access allowed with regards to a Google Drive object. This is a subset of the fields contained in a corresponding Drive Permissions object.
}
\seealso{
Other res: \code{\link{Activity}}, \code{\link{Event}},
  \code{\link{ListActivitiesResponse}}, \code{\link{Move}},
  \code{\link{Parent}}, \code{\link{PermissionChange}},
  \code{\link{Photo}}, \code{\link{Rename}},
  \code{\link{Target}}, \code{\link{User}}
}

