% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{PermissionChange}
\alias{PermissionChange}
\title{PermissionChange Object}
\usage{
PermissionChange(addedPermissions = NULL, removedPermissions = NULL)
}
\arguments{
\item{addedPermissions}{Lists all Permission objects added}

\item{removedPermissions}{Lists all Permission objects removed}
}
\value{
PermissionChange object
}
\description{
PermissionChange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a Drive object's permissions that changed as a result of a permissionChange type event.
}
