% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{Rename}
\alias{Rename}
\title{Rename Object}
\usage{
Rename(newTitle = NULL, oldTitle = NULL)
}
\arguments{
\item{newTitle}{The new title}

\item{oldTitle}{The old title}
}
\value{
Rename object
}
\description{
Rename Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a renametype event.
}
\seealso{
Other res: \code{\link{Activity}}, \code{\link{Event}},
  \code{\link{ListActivitiesResponse}}, \code{\link{Move}},
  \code{\link{Parent}}, \code{\link{PermissionChange}},
  \code{\link{Permission}}, \code{\link{Photo}},
  \code{\link{Target}}, \code{\link{User}}
}

