% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{Target}
\alias{Target}
\title{Target Object}
\usage{
Target(id = NULL, mimeType = NULL, name = NULL)
}
\arguments{
\item{id}{The ID of the target}

\item{mimeType}{The MIME type of the target}

\item{name}{The name of the target}
}
\value{
Target object
}
\description{
Target Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about the object modified by the event.
}
\seealso{
Other res: \code{\link{Activity}}, \code{\link{Event}},
  \code{\link{ListActivitiesResponse}}, \code{\link{Move}},
  \code{\link{Parent}}, \code{\link{PermissionChange}},
  \code{\link{Permission}}, \code{\link{Photo}},
  \code{\link{Rename}}, \code{\link{User}}
}

