% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{Photo}
\alias{Photo}
\title{Photo Object}
\usage{
Photo(url = NULL)
}
\arguments{
\item{url}{The URL of the photo}
}
\value{
Photo object
}
\description{
Photo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Photo information for a user.
}
