% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsactivity_objects.R
\name{User}
\alias{User}
\title{User Object}
\usage{
User(isDeleted = NULL, isMe = NULL, name = NULL, permissionId = NULL,
  photo = NULL)
}
\arguments{
\item{isDeleted}{A boolean which indicates whether the specified User was deleted}

\item{isMe}{Whether the user is the authenticated user}

\item{name}{The displayable name of the user}

\item{permissionId}{The permission ID associated with this user}

\item{photo}{The profile photo of the user}
}
\value{
User object
}
\description{
User Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a user.
}
